Cleanup - move  Attachment helper functions to Attachment.kt
