+++ core/src/main/kotlin/net/corda/core/contracts/Attachment.kt
@@ -2,7 +2,7 @@ package net.corda.core.contracts
 
 import net.corda.core.DoNotImplement
 import net.corda.core.identity.Party
-import net.corda.core.internal.contractVersion
+import net.corda.core.internal.entries
 import net.corda.core.internal.extractFile
 import net.corda.core.serialization.CordaSerializable
 import java.io.FileNotFoundException
@@ -50,6 +50,31 @@ interface Attachment : NamedByHash {
      * @throws [FileNotFoundException] if the given path doesn't exist in the attachment.
      */
     fun extractFile(path: String, outputTo: OutputStream) = openAsJAR().use { it.extractFile(path, outputTo) }
+
+    /**
+     * Checks whether the Attachment JAR contains a specific file entry.
+     * @param className The exact file name (including path, e.g. `"com/example/MyClass.class"`)
+     *                  to look for inside the JAR.
+     * @return `true` if an entry with the given name exists in the JAR, `false` otherwise.
+     * It is typically used to determine whether an attachment provides a given class or resource.
+     */
+    fun hasFile(className: String): Boolean = openAsJAR().use { it.entries().any { entry -> entry.name == className } }
+
+    /**
+     * @return `true` if the attachment is a JDK 17 compiled jar (class file major version 61), `false` otherwise.
+     */
+    @Suppress("MagicNumber")
+    fun isJdk17Jar(): Boolean = openAsJAR().use { jar ->
+        val firstClass = jar.entries().firstOrNull { it.name.endsWith(".class") }
+        if (firstClass != null) {
+            val header = ByteArray(8)
+            jar.read(header)
+            val major = ((header[6].toInt() and 0xFF) shl 8) or (header[7].toInt() and 0xFF)
+            major == 61
+        } else {
+            false
+        }
+    }
 
     /**
      * The parties that have correctly signed the whole attachment.
